refactor function home page

+++ module_home/lib/home_controller.dart
@@ -21,14 +21,20 @@ class HomeController extends ChangeNotifier implements ReassembleHandler {
   ChapterModel chapterSelected = ChapterModel();
   VerseModel verseSelected = VerseModel();
   List<VerseModel> versesList = [];
+  List<VersesMarkedModel> listMarkedModel = <VersesMarkedModel>[];
 
-  int? idVerseClicked;
+  int indexItemClicked = -1;
+  VersesMarkedModel verseIfExists = const VersesMarkedModel();
 
   Color pickerColor = const Color.fromARGB(255, 124, 20, 180);
   Color currentColor = const Color.fromARGB(255, 154, 14, 224);
 
-  void changeColor(Color color) {
+  Future<void> changeColor({required Color color}) async {
     pickerColor = color;
+  }
+
+  void setIndexVerseClicked(int index) {
+    indexItemClicked = index;
   }
 
   void setDefaultValuesBible({
@@ -56,13 +62,14 @@ class HomeController extends ChangeNotifier implements ReassembleHandler {
     }
   }
 
-  void setVerseSelected({required int index}) => verseSelected = versesList[index];
-  void changeVerseMarkedStatus({required int index}) {
+  void setVerseSelected({required int index}) {
     verseSelected = versesList[index];
-    bool oldValueIsMarked = !versesList[index].isMarked;
-    versesList[index] = versesList[index]
-        .copyWith(isMarked: oldValueIsMarked, colorMarked: pickerColor);
     notifyListeners();
+  }
+
+  void changeVerseMarkedStatus({required bool valueForMarked}) {
+    versesList[indexItemClicked] = versesList[indexItemClicked]
+        .copyWith(colorMarked: pickerColor, isMarked: valueForMarked);
   }
 
   Future<void> changeTheme() async {
@@ -72,29 +79,49 @@ class HomeController extends ChangeNotifier implements ReassembleHandler {
     notifyListeners();
   }
 
-  Future<void> addVerseMarkedOnTable(
-      {required VersesMarkedModel verseMarkedModel}) async {
-    idVerseClicked = await _localService.insertValues(
-        table: VersesMarkedTable.tableName, values: verseMarkedModel.toMap());
+  Future<void> addVerseMarkedOnTable({required Color color}) async {
+    if (verseIfExists.id == -1) {
+      VersesMarkedModel verseModelSelected = VersesMarkedModel(
+          bookId: bookSelected.id,
+          chapterId: chapterSelected.id,
+          verseId: verseSelected.id,
+          colorMarked: pickerColor);
+      changeColor(color: color);
+      changeVerseMarkedStatus(valueForMarked: true);
+      int idItemInserted = await _localService.insertValues(
+          table: VersesMarkedTable.tableName,
+          values: verseModelSelected.toMap());
+      verseIfExists = verseIfExists.copyWith(id: idItemInserted);
+      ///Set for true for leave icons of delete verse and add annotation
+      verseSelected = verseSelected.copyWith(isMarked: true);
+    } else {
+      changeColor(color: color);
+      changeVerseMarkedStatus(valueForMarked: true);
+      await updateColorVerseMarked(model: verseIfExists);
+    }
+    notifyListeners();
   }
 
-  Future<void> deleteVerseMarkedOnTable({required int id}) async {
-    await _localService.delete(
+  Future<void> updateColorVerseMarked(
+      {required VersesMarkedModel model}) async {
+    model = model.copyWith(colorMarked: pickerColor);
+    await _localService.updateValue(
         table: VersesMarkedTable.tableName,
-        whereSentence: '${VersesMarkedTable.id} = ? ',
-        whereArgs: [id.toString()]);
+        values: model.toMap(),
+        whereSentence: 'id = ?',
+        whereArgs: [model.id.toString()]);
   }
 
-  Future<int> alreadyVerseThisBase(VersesMarkedModel versesMarkedModel) async {
-    await getVersesMarkedOnTable();
-    VersesMarkedModel hasThisElementOnList = _appStore.listMarkedModel
-        .singleWhere(
-            (element) =>
-                element.bookId == versesMarkedModel.bookId &&
-                element.chapterId == versesMarkedModel.chapterId &&
-                element.verseId == versesMarkedModel.verseId,
-            orElse: () => const VersesMarkedModel(id: -1));
-    return hasThisElementOnList.id;
+  Future<void> deleteVerseMarkedOnTable() async {
+    if (verseIfExists.id != -1) {
+      await _localService.delete(
+          table: VersesMarkedTable.tableName,
+          whereSentence: '${VersesMarkedTable.id} = ? ',
+          whereArgs: [verseIfExists.id.toString()]);
+      changeVerseMarkedStatus(valueForMarked: false);
+      verseIfExists = verseIfExists.copyWith(id: -1);
+      notifyListeners();
+    }
   }
 
   Future<void> increaseFontSize() async {
@@ -118,12 +145,12 @@ class HomeController extends ChangeNotifier implements ReassembleHandler {
   Future<void> getVersesMarkedOnTable() async {
     List response =
         await _localService.getValues(table: VersesMarkedTable.tableName);
-    _appStore.listMarkedModel =
+    listMarkedModel =
         response.map((e) => VersesMarkedModel.fromMap(e)).toList();
     notifyListeners();
   }
 
-  Future<int> getIdVerseOnDatabase() async {
+  Future<void> getIdVerseOnDatabase() async {
     List<Map<String, Object?>> values = await _localService.getValues(
         table: VersesMarkedTable.tableName,
         whereSentence:
@@ -134,9 +161,12 @@ class HomeController extends ChangeNotifier implements ReassembleHandler {
           verseSelected.id.toString()
         ]);
     if (values.isNotEmpty) {
-      debugPrint(int.tryParse(values.first['id'].toString()).toString());
+      VersesMarkedModel versesMarkedModel =
+          VersesMarkedModel.fromMap(values.first);
+      verseIfExists = versesMarkedModel;
+    } else {
+      verseIfExists = const VersesMarkedModel(id: -1);
     }
-    return 1;
   }
 
   @override
